// src/entities/user/model/types.ts
export interface ICurrentStats {
	weeklyRank: number;
	activeGoals: number;
	activeLists: number;
	level: number;
	currentExperience: number;
	nextLevelExperience: number;
}

export interface ITotalStats {
	reviewsCount: number;
	completedGoals: number;
	completedLists: number;
	achievementsCount: number;
	weeklyCompletedChallenges: number;
	totalWeeklyChallenges: number;
}

export interface IWeeklyProgressItem {
	weekNumber: number;
	month: number;
	completedGoals: number;
	startDate: string;
	endDate: string;
}

export interface IStatGoal {
	completed: number;
	total: number;
}

export interface IHundredGoals {
	easy: IStatGoal;
	medium: IStatGoal;
	hard: IStatGoal;
}

export interface IUserStatistics {
	currentStats: ICurrentStats;
	totalStats: ITotalStats;
	weeklyProgress: Array<IWeeklyProgressItem>;
	hundredGoals: IHundredGoals;
}